ETL Tea, Project using Excel and SQL
This script should create a table called “business_unit” and a
table called “Product_BU.” Unfortunately, the metadata descriptions have been lost, so I need to figure out what you can
from the SQL script. The only thing you know about the metadata is that the company runs several individual strategic business
units, such as “On The Go” and “Snack.” Each of these business units is run under an umbrella designation, such as “Growth” or
“Decline.” The company will run marketing for growth products differently than it would run marketing for products on the decline.


You also have product order files from 2012, 2013, and 2014. They are attached as .csv files titled
 “2012_product_data_students.csv”
 “2013_product_data_students.csv”
 “2014_product_data_students.csv”
Your job is to use SQL to perform an ETL which will accomplish the following:
1. Extract data from the 2012, 2013, and 2014 order files
2. Transform the data according to the given rules
3. Load it into one final table
4. Export your final output table under the name “GX_output_final.csv” . (You may create as many or as few data objects as you
like in your work, but the data in the .csv file named “GX_output_final.csv” will be the data evaluated.